// frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        SystemUnderTestModel: {
            /** Sut Id */
            sut_id?: string | null;
            /** System Name */
            system_name: string;
            /** Description */
            description?: string | null;
            /** Wiki Url */
            wiki_url?: string | null;
            /** Account Id */
            account_id: string;
            /** Owner User Id */
            owner_user_id: string;
            /**
             * Is Active
             * @default true
             */
            is_active: boolean;
            /** Deactivated At */
            deactivated_at?: string | null;
            /** Deactivated By User Id */
            deactivated_by_user_id?: string | null;
            /**
             * Created At
             * Format: date-time
             * @default 2026-01-15T20:10:27.663069Z
             */
            created_at: string;
            /** Updated At */
            updated_at?: string | null;
        };
// ...
        TestCaseModel: {
            /** Test Case Id */
            test_case_id?: string | null;
            /** Test Name */
            test_name: string;
            /** Description */
            description?: string | null;
            /**
             * Test Type
             * @default functional
             */
            test_type: string;
            /** Sut Id */
            sut_id: string;
            /** Owner User Id */
            owner_user_id: string;
            /** Account Id */
            account_id: string;
            /**
             * Is Active
             * @default true
             */
            is_active: boolean;
            /** Deactivated At */
            deactivated_at?: string | null;
            /** Deactivated By User Id */
            deactivated_by_user_id?: string | null;
            /**
             * Created At
             * Format: date-time
             * @default 2026-01-15T20:10:27.663690Z
             */
            created_at: string;
            /** Updated At */
            updated_at?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}